finance-microservice/requirements.txt
# Web Framework
fastapi>=0.95.0
uvicorn>=0.20.0
python-multipart>=0.0.5
httpx>=0.23.0

# Datenbank
sqlalchemy>=2.0.0
alembic>=1.10.0
asyncpg>=0.27.0
psycopg2-binary>=2.9.5
databases>=0.7.0

# Caching und Messaging
redis>=4.5.4
aioredis>=2.0.0
msgpack>=1.0.5

# LLM-Integration
langchain>=0.0.200
langchain-openai>=0.0.2
langchain-anthropic>=0.0.7
llama-index>=0.8.0
openai>=0.27.0
anthropic>=0.5.0
tiktoken>=0.4.0
sentence-transformers>=2.2.2

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.2
typer>=0.9.0
rich>=13.3.5
structlog>=23.1.0
prometheus-client>=0.16.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
pytest-mock>=3.10.0

# Type-Checking und Linting
mypy>=1.2.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
pre-commit>=3.2.2 